smart-bin-interface/requirements.txt
streamlit>=1.28.0
opencv-python>=4.8.0
qrcode[pil]>=7.4.2
Pillow>=10.0.0
numpy>=1.24.0

# Optional ML libraries for better image recognition
torch>=1.9.0
torchvision>=0.10.0
transformers>=4.20.0
requests>=2.28.0

